+++ src/config/webpack.serve.babel.js
@@ -4,12 +4,10 @@
  * @module config/webpack.serve.babel
  */
 
-// import fs from 'fs';
 import path from 'path';
 import { isString, isArray, isObject, isFunction } from 'util';
 import webpack from 'webpack';
 import OpenBrowserPlugin from 'open-browser-webpack-plugin';
-// import importFresh from 'import-fresh';
 import pRequire from '../util/require';
 
 const { NODE_ENV } = process.env;
@@ -59,10 +57,9 @@ const pushClientJS = (entry) => {
  */
 const webpackConfig = (program) => {
   const { CONTEXT } = process.env;
-  const projectRootPath = CONTEXT ? path.resolve(CONTEXT) : process.cwd();
-  const assetsPath = path.resolve(projectRootPath, assetsDist);
-  const dllPath = path.resolve(projectRootPath, dll);
-  const context = projectRootPath;
+  const context = CONTEXT ? path.resolve(CONTEXT) : process.cwd();
+  const assetsPath = path.resolve(context, assetsDist);
+  const dllPath = path.resolve(context, dll);
 
   let entry = isFunction(entries) ? entries() : entries;
 
@@ -141,37 +138,6 @@ const webpackConfig = (program) => {
       }
     })
   ];
-
-  // 添加 html 模版
-  // 先只考虑 object 类型的 entries
-  // Object.keys(entry).forEach((chunkName) => {
-  //   const globalMockFile = path.resolve(projectRootPath, 'mock/pages/__global.js');
-  //   const mockFile = path.resolve(projectRootPath, `mock/pages/${chunkName}.js`);
-  //   let globalMockData = {};
-  //   let mockData = {};
-  //   if (fs.existsSync(globalMockFile)) {
-  //     // globalMockData = importFresh(globalMockFile);
-  //     globalMockData = require(globalMockFile); // eslint-disable-line
-  //     if (globalMockData.default) {
-  //       globalMockData = globalMockData.default;
-  //     }
-  //   }
-  //   if (fs.existsSync(mockFile)) {
-  //     // mockData = importFresh(mockFile);
-  //     mockData = require(mockFile); // eslint-disable-line
-  //     if (mockData.default) {
-  //       mockData = mockData.default;
-  //     }
-  //   }
-  //   const mock = { ...mockData, ...globalMockData };
-  //   plugins.push(new HtmlWebpackPlugin({
-  //     filename: `${chunkName}.html`,
-  //     template: 'template.html',
-  //     title: chunkName,
-  //     chunks: Object.keys(commonChunks || {}).concat(chunkName),
-  //     mock
-  //   }));
-  // });
 
   if (hot) {
     entry = pushClientJS(entry);